backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2

# Data & ML
yfinance==0.2.44
pandas==2.2.3
numpy==1.26.4
pandas-ta==0.3.14b
scikit-learn==1.5.2
scipy==1.14.1
shap==0.46.0
lightgbm==4.5.0
catboost==1.2.7
mlflow-skinny==2.18.0
joblib==1.4.2
python-dateutil==2.9.0

# NLP / Sentiment
vaderSentiment==3.3.2

# Task queue
celery==5.4.0
redis==5.1.1

xgboost==2.1.1

# Utils
httpx==0.27.2
python-dotenv==1.0.1
structlog==24.4.0
slowapi==0.1.9

# Auth
pyjwt==2.9.0
bcrypt==4.2.0

# Notifications
aiosmtplib==3.0.2

# Tests
pytest==8.3.3
pytest-asyncio==0.24.0
